+++ src/vm/hints/pow_utils.rs
@@ -14,7 +14,7 @@ Implements hint:
 */
 pub fn pow(
     vm_proxy: &mut VMProxy,
-    ids: &HashMap<String, BigInt>,
+    ids: &HashMap<String, usize>,
     hint_ap_tracking: Option<&ApTracking>,
 ) -> Result<(), VirtualMachineError> {
     let prev_locs_addr =
